+++ psbts/finalize_psbt.js
@@ -2,6 +2,7 @@ const updatePsbt = require('./update_psbt');
 
 /** Finalize the inputs of a PSBT
   {
+    ecp: <ECPair Object>
     psbt: <BIP 174 Encoded PSBT Hex String>
   }
 
@@ -13,6 +14,6 @@ const updatePsbt = require('./update_psbt');
     psbt: <BIP 174 Encoded PSBT Hex String>
   }
 */
-module.exports = ({psbt}) => {
-  return updatePsbt({is_final: true, psbt});
+module.exports = ({ecp, psbt}) => {
+  return updatePsbt({ecp, psbt, is_final: true});
 };