+++ test/psbts/test_decode_psbt.js
@@ -1,4 +1,5 @@
 const {test} = require('tap');
+const tinysecp = require('tiny-secp256k1');
 
 const {decodePsbt} = require('./../../');
 
@@ -996,7 +997,9 @@ const tests = {
 
 // Run the tests
 Object.keys(tests).map(t => tests[t]).forEach(({args, err, msg, result}) => {
-  return test(msg, ({end, equal, ok, throws}) => {
+  return test(msg, async ({end, equal, ok, throws}) => {
+    args.ecp = (await import('ecpair')).ECPairFactory(tinysecp);
+
     if (!!err) {
       throws(() => decodePsbt(args), new Error(err));