+++ test/psbts/test_extract_transaction.js
@@ -1,4 +1,5 @@
 const {test} = require('tap');
+const tinysecp = require('tiny-secp256k1');
 
 const {decodePsbt} = require('./../../');
 const {extractTransaction} = require('./../../');
@@ -18,7 +19,11 @@ const tests = {
 
 // Run the tests
 Object.keys(tests).map(t => tests[t]).forEach(({args, err, msg, result}) => {
-  return test(msg, ({end, equal}) => {
+  return test(msg, async ({end, equal}) => {
+    const ecp = (await import('ecpair')).ECPairFactory(tinysecp);
+
+    args.ecp = ecp;
+
     const {transaction} = extractTransaction(args);
 
     equal(transaction, result.transaction);