+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Versions
 
+## 2.0.0
+
+- Update bitcoinjs-lib dependency
+
+### Breaking Changes
+
+- `combinePsbts`, `decodePsbt`, `extractTransaction`, `finalizePsbt`,
+    `signPsbt`, `transactionAsPsbt`, and `updatePsbt` now require `ecp` object
+
 ## 1.1.11
 
 - `updatePsbt`: Add support for bip32 derivation paths on p2wpkh spends